Add files via upload
Full page of classes for use in the system. I tried to make them all properly scalable within each other through the use of foreign keys so that we can implement all required information. They are not currently populated but we can populate these classes and then draw from them as we see fit