Populate database further. Add button generation
I added more users and courses to the database. I also signed a few people up for classes to test the button functionality. System will generate buttons for every class a user is signed up for. The buttons don't work and the system cannot figure out who is logged in. It's a super easy fix for someone who has that global variable running properly on their computer though