populate databases, upload functional buttons system
I might have accidentally pushed unchanged pages to the repo last time. This should have the code that generates buttons on the dashboard. It links to a blank page but it does link to something now and I added the buttons into the css so they match the rest of the page. I think we should have them boxed off in a div with a bold outline but I didn't bother doing that yet. For now it's all uniform